+++ main/ArduinoBridge.h
@@ -7,17 +7,19 @@
 
 #ifndef __ARDUINO_BRIDGE
 #define __ARDUINO_BRIDGE
+
+#include <cstddef>
 #include "Pointer.h"
 #include "Packet.h"
 #include "Callsign.h"
 
-unsigned long int arduino_millis();
-long int arduino_random(long int min, long int max);
+uint32_t arduino_millis();
+int32_t arduino_random(int32_t min, int32_t max);
 void logs(const char*, const char*);
-void logi(const char*, long int);
+void logi(const char*, int32_t);
 void app_recv(Ptr<Packet>);
-unsigned int arduino_nvram_id_load();
-void arduino_nvram_id_save(unsigned int);
+uint32_t arduino_nvram_id_load();
+void arduino_nvram_id_save(uint32_t);
 Callsign arduino_nvram_callsign_load();
 void arduino_nvram_callsign_save(const Callsign&);
 Buffer arduino_nvram_load(const char *);
